Add helpful functions to Span<T> and a lazy test
begin() and end() allow the use of range based for loops with the span. Added in front() and back() since that can be useful at times, and operator[] since it's pretty frequently used in my experience. Added a simple test of these functions to main, using the VS debugger to verify proper behavior because I'm being lazy.